gestion utilisateur et droits terminé

+++ app/Http/Controllers/ProfileController.php
@@ -9,27 +9,59 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\View\View;
+use App\Models\Role;
 
 class ProfileController extends Controller
 {
-    public function index(): View
+    public function index(Request $request): View
     {
-        $users = User::all();
-        return view(
-            'profile.index',
-            [
-                'users' => $users
-            ]
-        );
+        $search = $request->input('search');
+        $show_deleted = $request->input('show_deleted');
+        if ($show_deleted) {
+            $users = User::onlyTrashed()->get();
+            return view(
+                'profile.index',
+                [
+                    'users' => $users
+                ]
+            );
+        } else {
+            // Rechercher des utilisateurs en fonction du terme de recherche (si fourni)
+            $users = User::query()
+                ->when($search, function ($query, $search) {
+                    $query->where('first_name', 'like', "%{$search}%")
+                        ->orWhere('last_name', 'like', "%{$search}%")
+                        ->orWhere('email', 'like', "%{$search}%")
+                        ->orWhere('phone', 'like', "%{$search}%")
+                        ->orWhereHas('role', function ($query) use ($search) {
+                            $query->where('name', 'like', "%{$search}%");
+                        });
+                })
+                ->with('role') // Ensure the role relationship is loaded
+                ->get();
+
+            return view(
+                'profile.index',
+                [
+                    'users' => $users
+                ]
+            );
+        }
     }
     /**
      * Display the user's profile form.
      */
     public function edit(int $id): View
     {
-        $user = User::find($id);
+
+        $user = User::findOrFail($id);
+        if (Auth::user()->role_id != 1 && $user->id != Auth::id()) {
+            abort(404);
+        }
+        $roles = Role::all();
         return view('profile.edit', [
             'user' => $user,
+            'roles' => $roles,
         ]);
     }
 
@@ -38,36 +70,33 @@ class ProfileController extends Controller
      */
     public function update(ProfileUpdateRequest $request): RedirectResponse
     {
-        $user = User::find($request->id);
-        $user->fill($request->only(['first_name', 'last_name', 'phone', 'email']));
+        /** @var \App\Models\User $user */
+        $user = User::findOrFail($request->id);
+        $user->update($request->only(['first_name', 'last_name', 'phone', 'email']));
 
-        // if ($user->isDirty('email')) {
-        //     $user->email_verified_at = null;
-        // }
+        return Redirect::route('profile.edit', ['id' => $user->id])->with('status', 'profile modifié');
+    }
+    public function updateAdmin(Request $request): RedirectResponse
+    {
+        /** @var \App\Models\User $user */
+        $user = User::findOrFail($request->id);
+        $user->update($request->only(['role_id']));
 
-        $user->save();
-
-        return Redirect::route('profile.edit', ['id'=> $user->id])->with('status', 'profile-updated');
+        return Redirect::route('profile.edit', ['id' => $user->id])->with('status', 'profile modifié');
     }
 
     /**
      * Delete the user's account.
      */
-    public function destroy(Request $request): RedirectResponse
+    public function destroy(User $user): RedirectResponse
     {
-        $request->validateWithBag('userDeletion', [
-            'password' => ['required', 'current_password'],
-        ]);
-
-        $user = $request->user();
-
-        Auth::logout();
-
-        $user->delete();
-
-        $request->session()->invalidate();
-        $request->session()->regenerateToken();
-
-        return Redirect::to('/');
+        if ($user->deleted_at === null) {
+            $user->delete();
+            \Log::info("User {$user->id} deleted.");
+            return Redirect::route('profile.index')->with('status', "compte $user->first_name $user->last_name supprimé");
+        } else {
+            $user->restore();
+            return Redirect::route('profile.edit', ['id' => $user->id])->with('status', 'compte restauré');
+        }
     }
 }
